+++ packages/backend/src/nest/const.ts
@@ -8,6 +8,14 @@ export enum Config {
   IPFS_REPO_PATH = 'Ipfs',
 }
 
+/**
+ * WARNING! We may automatically delete directories and files listed
+ * in this configuration enum after certain tests. So you may want to
+ * be careful when changing any of these values. For example, you
+ * probably do not want to accidentally set ORBIT_DB_DIR to `/home/my`
+ * or `/home/my/favorite/files` or anything like that, otherwise those
+ * directories may be deleted.
+ */
 export enum TestConfig {
   QUIET_DIR = '.quiet-test-nest-backend',
   PEER_ID_FILENAME = 'peerIdKey-test-nest-backend',