+++ packages/backend/src/nest/local-db/local-db.types.ts
@@ -3,5 +3,6 @@ export enum LocalDBKeys {
   REGISTRAR = 'registrar',
   PEERS = 'peers',
   PSK = 'psk',
+  OWNER_ORBIT_DB_IDENTITY = 'ownerOrbitDbIdentity',
 }
 export type LocalDbStatus = 'opening' | 'open' | 'closing' | 'closed'