+++ packages/backend/src/nest/registration/registration.functions.ts
@@ -64,6 +64,7 @@ export const extractPendingCsrs = async (payload: { csrs: string[]; certificates
       pendingCsrs.push(csr)
     }
   }
+  logger('DuplicatedCertBug', { parsedUniqueCsrs, pendingNames, certNames })
   return pendingCsrs
 }