+++ packages/backend/src/nest/storage/storage.service.spec.ts
@@ -11,12 +11,10 @@ import {
 import {
   ChannelMessage,
   Community,
-  ConnectionProcessInfo,
   FileMetadata,
   Identity,
   MessageType,
   PublicChannel,
-  SocketActionTypes,
   TestMessage,
 } from '@quiet/types'
 
@@ -33,7 +31,6 @@ import { Libp2pService } from '../libp2p/libp2p.service'
 import { SocketModule } from '../socket/socket.module'
 import { StorageModule } from './storage.module'
 import { StorageService } from './storage.service'
-import { StorageEvents } from './storage.types'
 import fs from 'fs'
 import { type FactoryGirl } from 'factory-girl'
 import { fileURLToPath } from 'url'
@@ -42,7 +39,10 @@ import { LocalDbModule } from '../local-db/local-db.module'
 import { LocalDbService } from '../local-db/local-db.service'
 import { IPFS_REPO_PATCH, ORBIT_DB_DIR, QUIET_DIR } from '../const'
 import { LocalDBKeys } from '../local-db/local-db.types'
-import { RegistrationEvents } from '../registration/registration.types'
+import { CertificatesRequestsStore } from './certifacteRequests/certificatesRequestsStore'
+import { CertificatesStore } from './certificates/certificates.store'
+import { CommunityMetadataStore } from './communityMetadata/communityMetadata.store'
+import { OrbitDb } from './orbitDb/orbitDb.service'
 
 const filename = fileURLToPath(import.meta.url)
 const dirname = path.dirname(filename)
@@ -69,6 +69,10 @@ describe('StorageService', () => {
   let libp2pService: Libp2pService
   let lazyModuleLoader: LazyModuleLoader
   let localDbService: LocalDbService
+  let certificatesRequestsStore: CertificatesRequestsStore
+  let certificatesStore: CertificatesStore
+  let communityMetadataStore: CommunityMetadataStore
+  let orbitDbService: OrbitDb
   let peerId: PeerId
 
   let store: Store
@@ -126,7 +130,12 @@ describe('StorageService', () => {
 
     storageService = await module.resolve(StorageService)
     localDbService = await module.resolve(LocalDbService)
+    orbitDbService = await module.resolve(OrbitDb)
 
+    certificatesRequestsStore = await module.resolve(CertificatesRequestsStore)
+    certificatesStore = await module.resolve(CertificatesStore)
+    communityMetadataStore = await module.resolve(CommunityMetadataStore)
+    console.log({ communityMetadataStore })
     lazyModuleLoader = await module.resolve(LazyModuleLoader)
 
     orbitDbDir = await module.resolve(ORBIT_DB_DIR)
@@ -257,10 +266,9 @@ describe('StorageService', () => {
       expect(db).not.toBe(undefined)
       if (!db) return // TS complaining
       const channelsDbAddress = storageService.channels?.address
-      const certificatesDbAddress = storageService.certificatesStore.getAddress()
-      const certificatesRequestsDbAddress = storageService.certificatesRequestsStore.getAddress()
-      // @ts-expect-error 'communityMetadata' is private
-      const communityMetadataDbAddress = storageService.communityMetadata.address
+      const certificatesDbAddress = certificatesStore.getAddress()
+      const certificatesRequestsDbAddress = certificatesRequestsStore.getAddress()
+      const communityMetadataDbAddress = communityMetadataStore.getAddress()
       expect(channelsDbAddress).not.toBeFalsy()
       expect(certificatesDbAddress).not.toBeFalsy()
       expect(subscribeToPubSubSpy).toBeCalledTimes(2)
@@ -276,7 +284,7 @@ describe('StorageService', () => {
     })
   })
 
-  describe.only('Certificate', () => {
+  describe('Certificate', () => {
     // FIXME: Due to moving certificates to a separate store and lack of proper nest configuration, this test is broken
     it.skip('username check fails if username is already in use', async () => {
       const userCertificate = await createUserCert(