+++ packages/mobile/src/store/init/deepLink/deepLink.saga.ts
@@ -24,6 +24,9 @@ export function* deepLinkSaga(action: PayloadAction<ReturnType<typeof initAction
 
   console.log('INIT_NAVIGATION: Continuing on deep link flow.')
 
+  // Reset deep link flag for future redirections sake
+  yield* put(initActions.resetDeepLink())
+
   const community = yield* select(communities.selectors.currentCommunity)
 
   // Link opened mid registration
@@ -80,6 +83,7 @@ export function* deepLinkSaga(action: PayloadAction<ReturnType<typeof initAction
     ownership: CommunityOwnership.User,
     peers: data.pairs,
     psk: data.psk,
+    ownerOrbitDbIdentity: data.ownerOrbitDbIdentity,
   }
 
   yield* put(communities.actions.createNetwork(payload))