+++ packages/state-manager/src/sagas/socket/startConnection/startConnection.saga.ts
@@ -38,12 +38,11 @@ import {
   type RemoveDownloadStatus,
   type SendCertificatesResponse,
   type SetChannelSubscribedPayload,
-  SocketActionTypes,
   type SavedOwnerCertificatePayload,
   type SendOwnerCertificatePayload,
-  CommunityMetadata,
-  SendCsrsResponse,
-  ConnectionProcessInfo,
+  type SendCsrsResponse,
+  type CommunityMetadata,
+  SocketActionTypes,
 } from '@quiet/types'
 
 const log = logger('socket')
@@ -70,12 +69,12 @@ export function subscribe(socket: Socket) {
     | ReturnType<typeof identityActions.throwIdentityError>
     | ReturnType<typeof identityActions.saveOwnerCertToDb>
     | ReturnType<typeof identityActions.savedOwnerCertificate>
+    | ReturnType<typeof identityActions.checkLocalCsr>
     | ReturnType<typeof communitiesActions.storePeerList>
     | ReturnType<typeof communitiesActions.updateCommunity>
     | ReturnType<typeof communitiesActions.responseRegistrar>
     | ReturnType<typeof communitiesActions.launchRegistrar>
     | ReturnType<typeof communitiesActions.launchCommunity>
-    | ReturnType<typeof communitiesActions.addOwnerCertificate>
     | ReturnType<typeof networkActions.addInitializedCommunity>
     | ReturnType<typeof networkActions.addInitializedRegistrar>
     | ReturnType<typeof networkActions.removeConnectedPeer>
@@ -169,6 +168,7 @@ export function subscribe(socket: Socket) {
       console.log('on SocketActionTypes.NEW_COMMUNITY')
       emit(identityActions.saveOwnerCertToDb())
       emit(publicChannelsActions.createGeneralChannel())
+      emit(identityActions.saveUserCsr())
     })
     socket.on(SocketActionTypes.REGISTRAR, (payload: ResponseRegistrarPayload) => {
       console.log('SocketActionTypes.REGISTRAR')
@@ -186,8 +186,6 @@ export function subscribe(socket: Socket) {
     socket.on(SocketActionTypes.COMMUNITY, (payload: ResponseLaunchCommunityPayload) => {
       console.log('Hunting for heisenbug: Community event received in state-manager')
       emit(communitiesActions.launchRegistrar(payload.id))
-      // Not sure about this saveUserCsr. It seems that we've added it to secure case when user closes the app unexpectedly before csr is saved to db, so we'll do that on restart.
-      emit(identityActions.saveUserCsr())
       emit(filesActions.checkForMissingFiles(payload.id))
       emit(networkActions.addInitializedCommunity(payload.id))
       emit(communitiesActions.clearInvitationCodes())
@@ -196,11 +194,18 @@ export function subscribe(socket: Socket) {
     })
     // Errors
     socket.on(SocketActionTypes.ERROR, (payload: ErrorPayload) => {
-      log(payload)
+      // FIXME: It doesn't look like log errors have the red error
+      // color in the console, which makes them difficult to find.
+      // Also when only printing the payload, the full trace is not
+      // available.
+      log.error(payload)
+      console.error(payload, payload.trace)
       emit(errorsActions.handleError(payload))
     })
     // Certificates
     socket.on(SocketActionTypes.RESPONSE_GET_CSRS, (payload: SendCsrsResponse) => {
+      console.log('REPONSE_GET_CSRS')
+      emit(identityActions.checkLocalCsr(payload))
       emit(usersActions.storeCsrs(payload))
     })
     socket.on(SocketActionTypes.RESPONSE_GET_CERTIFICATES, (payload: SendCertificatesResponse) => {
@@ -208,13 +213,6 @@ export function subscribe(socket: Socket) {
     })
     socket.on(SocketActionTypes.SEND_USER_CERTIFICATE, (payload: SendOwnerCertificatePayload) => {
       log(`${SocketActionTypes.SEND_USER_CERTIFICATE}: ${payload.communityId}`)
-
-      emit(
-        communitiesActions.addOwnerCertificate({
-          communityId: payload.communityId,
-          ownerCertificate: payload.payload.ownerCert,
-        })
-      )
 
       emit(
         communitiesActions.storePeerList({
@@ -232,6 +230,7 @@ export function subscribe(socket: Socket) {
         communitiesActions.updateCommunity({
           id: payload.communityId,
           rootCa: payload.payload.rootCa,
+          ownerCertificate: payload.payload.ownerCert,
         })
       )
       emit(communitiesActions.launchCommunity(payload.communityId))
@@ -239,8 +238,8 @@ export function subscribe(socket: Socket) {
     socket.on(SocketActionTypes.SAVED_OWNER_CERTIFICATE, (payload: SavedOwnerCertificatePayload) => {
       log(`${SocketActionTypes.SAVED_OWNER_CERTIFICATE}: ${payload.communityId}`)
       emit(
-        communitiesActions.addOwnerCertificate({
-          communityId: payload.communityId,
+        communitiesActions.updateCommunity({
+          id: payload.communityId,
           ownerCertificate: payload.network.certificate,
         })
       )
@@ -252,14 +251,9 @@ export function subscribe(socket: Socket) {
       )
       emit(identityActions.savedOwnerCertificate(payload.communityId))
     })
-    socket.on(SocketActionTypes.SAVE_COMMUNITY_METADATA, (payload: CommunityMetadata) => {
-      log(`${SocketActionTypes.SAVE_COMMUNITY_METADATA}: ${payload}`)
-      emit(
-        communitiesActions.saveCommunityMetadata({
-          rootCa: payload.rootCa,
-          ownerCertificate: payload.ownerCertificate,
-        })
-      )
+    socket.on(SocketActionTypes.COMMUNITY_METADATA_SAVED, (payload: CommunityMetadata) => {
+      log(`${SocketActionTypes.COMMUNITY_METADATA_SAVED}: ${payload}`)
+      emit(communitiesActions.saveCommunityMetadata(payload))
     })
     socket.on(SocketActionTypes.LIBP2P_PSK_SAVED, (payload: { psk: string }) => {
       log(`${SocketActionTypes.LIBP2P_PSK_SAVED}`)